fixed: compile problems
git-svn-id: https://external.lite3d.com/opensg/svn_repo/trunk@2402 4683daeb-ad0f-0410-a623-93161e962ae5

+++ Source/System/RenderingBackend/OSGOcclusionCullingTreeBuilder.cpp
@@ -891,7 +891,7 @@ OcclusionCullingTreeBuilder::addNodeAdaptiveBucketMode(OCRenderTreeNode *pNode)
     bucketIdx = osgClamp<Int32>(0, bucketIdx, _numBuckets - 1);
 
     // XXX with the above osgClamp this never triggers... ?? --cneumann
-    if(bucketIdx < 0 || bucketIdx >= _numBuckets)
+    if(bucketIdx < 0 || bucketIdx >= Int32(_numBuckets))
     {
         Real32 newLow   = osgMin(val, _bucketLow );
         Real32 newHigh  = osgMax(val, _bucketHigh);
